REF: de-duplicate get_indexer methods (#38372)
